% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lag_distances.R
\name{calc_lag_distances}
\alias{calc_lag_distances}
\title{Calculate distance between consecutive points}
\usage{
calc_lag_distances(lats, lons)
}
\arguments{
\item{lats}{A numeric vector of latitude values.}

\item{lons}{A numeric vector of longitude values.}
}
\value{
A numeric vector of distances in meters between consecutive pairs of
points. Preceeded by a \code{NA}.
}
\description{
Calculate distance between consecutive points
}
