% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{get_type_names_mapping}
\alias{get_type_names_mapping}
\title{Extract ship type to names mapping}
\usage{
get_type_names_mapping(ships_data_path)
}
\arguments{
\item{ships_data_path}{A path to ships.csv input data.}
}
\value{
A named \code{list} with ship types as names and ship names as elements.
}
\description{
Extract ship type to names mapping
}
