% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_rows}
\alias{get_unique_rows}
\title{Get unique rows from a \code{data.frame}}
\usage{
get_unique_rows(dt, col_names)
}
\arguments{
\item{dt}{An input \code{data.frame}.}

\item{col_names}{A vector of column names to subset from \code{df}.}
}
\value{
A \code{data.frame} with unique rows.
}
\description{
Get unique rows from a \code{data.frame}
}
