% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{get_unique_type_name_pairs}
\alias{get_unique_type_name_pairs}
\title{Extract unique ship type and name pairs}
\usage{
get_unique_type_name_pairs(ships_data_path)
}
\arguments{
\item{ships_data_path}{A path to ships.csv input data.}
}
\value{
A \code{data.table} with unique (ship type, ship name) pairs.
}
\description{
Extract unique ship type and name pairs
}
