% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ships}
\alias{ships}
\title{Longest distance that ships traveled.}
\format{
A data frame with 2284 rows and 5 columns.

\describe{
  \item{lat}{latitude of ship's location}
  \item{lon}{longitude of ship's location}
  \item{ship_type}{the type of ship}
  \item{ship_name}{the ship's name}
  \item{max_distance}{longest distanced traveled by the ship}
}
}
\usage{
ships
}
\description{
A data set containing information about the longest distance ships traveled.
There are two rows of data for each ship with different coordinates
 indicating the positions for which the ship traveled the longest distance.
}
\keyword{datasets}
