% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{get_ships_distances}
\alias{get_ships_distances}
\title{Extract max distance per ship between consecutive observations}
\usage{
get_ships_distances(ships_data_path)
}
\arguments{
\item{ships_data_path}{A path to ships.csv input data.}
}
\value{
A \code{data.table} with SHIP_ID, travel start point, end point, and
travelled distance for the maximum travelled distance.
}
\description{
Extract max distance per ship between consecutive observations
}
