% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{get_type_names_mapping}
\alias{get_type_names_mapping}
\title{Extract ship type to names mapping}
\usage{
get_type_names_mapping(ships)
}
\arguments{
\item{ships}{A \code{data.table} with ships raw data.}
}
\value{
A named \code{list} with ship types as names and ship names as elements.
}
\description{
Extract ship type to names mapping
}
